+++ api/src/controllers/user.controller.ts
@@ -1,12 +1,24 @@
 import { Request, Response, NextFunction } from 'express'
+import jwt from 'jsonwebtoken'
 
 import User from '../models/User'
 import userService from '../services/user.service'
-import { Slug } from '../@types/common'
-import { BadRequestError, ForbiddenError } from '../helpers/apiError'
+import { Slug, Token } from '../@types/common'
+import {
+  BadRequestError,
+  ForbiddenError,
+  InternalServerError,
+  NotFoundError,
+} from '../helpers/apiError'
 import { sendResponse } from '../helpers/responseHandler'
-import { User as UserType, UserQuery, UserUpdateFields } from '../@types/user'
+import {
+  User as UserType,
+  UserDocument,
+  UserQuery,
+  UserUpdateFields,
+} from '../@types/user'
 import mongoose from 'mongoose'
+import { JWT_SECRET } from '../util/secrets'
 
 export const signUp = async (
   req: Request<{}, {}, UserType, {}>,
@@ -24,6 +36,7 @@ export const signUp = async (
     const token = await userService.signUp(user)
 
     sendResponse(res, {
+      status: 'success',
       statusCode: 201,
       message: `Registered a new user '${user.name}'. Please, verify your email address to finish registration.`,
       payload: { token },
@@ -45,12 +58,23 @@ export const verify = async (
   try {
     const { token } = req.query
 
-    const verifiedUser = await userService.verify(token)
+    jwt.verify(token, JWT_SECRET, async (error: any, decoded: any) => {
+      if (error) {
+        throw new BadRequestError('Token has expired.')
+      }
+      const verifiedUser = await userService.verify(decoded as Token)
 
-    sendResponse(res, {
-      statusCode: 201,
-      message: `Verified the user '${verifiedUser.name}'. Welcome to Pierre's General Store`,
-      payload: verifiedUser,
+      if (!verifiedUser)
+        throw new InternalServerError(
+          'Could not verify this user. Try again later.'
+        )
+
+      sendResponse(res, {
+        status: 'success',
+        statusCode: 201,
+        message: `Verified the user '${verifiedUser.name}'. Welcome to Pierre's General Store`,
+        payload: verifiedUser,
+      })
     })
   } catch (error) {
     if (error instanceof Error && error.name == 'ValidationError') {
@@ -73,7 +97,11 @@ export const getUser = async (
       ? await userService.findById(id)
       : await userService.findAll()
 
+    if (!foundUsers) {
+      throw new NotFoundError(`User with the ID: '${id}' is not found`)
+    }
     sendResponse(res, {
+      status: 'success',
       statusCode: 200,
       message: 'Returned users',
       payload: foundUsers,
@@ -103,7 +131,13 @@ export const updateUser = async (
 
     const updatedUser = await userService.update(id, update)
 
+    if (!updatedUser) {
+      throw new InternalServerError(
+        `Could not update a user with ID: '${id}'. Try again later.`
+      )
+    }
     sendResponse(res, {
+      status: 'success',
       statusCode: 200,
       message: 'Successfully updated a user',
       payload: updatedUser,
@@ -125,9 +159,20 @@ export const deleteUser = async (
   try {
     const { id } = req.query
 
-    await userService.remove(id)
+    const foundUser = await User.findById(id)
+
+    if (!foundUser) throw new NotFoundError(`A user '${id}' does not exist.`)
+
+    const removedUser = await userService.remove(id)
+
+    if (removedUser.deletedCount === 0) {
+      throw new InternalServerError(
+        `Could not delete a user '${id}'. Try again later.`
+      )
+    }
 
     sendResponse(res, {
+      status: 'success',
       statusCode: 200,
       message: `Deleted a user '${id}'`,
     })
@@ -153,6 +198,7 @@ export const getProfile = async (
     const user = await userService.findById(id)
 
     sendResponse(res, {
+      status: 'success',
       statusCode: 200,
       message: 'Profile of the user',
       payload: user,
